:recycle: refactor(TEC-38): [Audsat] Considerações de codereview
