+++ src/main/java/br/com/audsat/carinsurancequotes/builders/InsuranceCalculatorBuilder.java
@@ -1,11 +1,11 @@
-package br.com.audsat.carinsurancequotes.adapters;
+package br.com.audsat.carinsurancequotes.builders;
 
 import br.com.audsat.carinsurancequotes.configurations.business.InsuranceRiskCalculatorConfiguration;
 import br.com.audsat.carinsurancequotes.domains.entities.CarEntity;
 import br.com.audsat.carinsurancequotes.domains.entities.DriverEntity;
 import br.com.audsat.carinsurancequotes.domains.pojo.InsuranceRiskCalculator;
 
-public interface InsuranceCalculatorAdapter {
+public interface InsuranceCalculatorBuilder {
     static InsuranceRiskCalculator build(final InsuranceRiskCalculatorConfiguration insuranceRiskCalculatorConfiguration,
                                          final DriverEntity mainDriverEntity,
                                          final DriverEntity driverEntity,
