+++ src/main/java/br/com/audsat/carinsurancequotes/domains/entities/CarEntity.java
@@ -1,12 +1,11 @@
 package br.com.audsat.carinsurancequotes.domains.entities;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
+import java.util.List;
 
 @Getter
 @NoArgsConstructor
@@ -28,5 +27,8 @@ public class CarEntity {
 
     @Column(name = "fipe_value")
     private BigDecimal fipeValue;
+
+    @OneToMany(mappedBy = "car")
+    private List<InsuranceEntity> insurances;
 
 }
